Refs #35948 - DRYed up the context export controllers
